+++ frontend/src/app/modules/shared/service/parts.service.ts
@@ -25,7 +25,6 @@ import { ApiService } from '@core/api/api.service';
 import { Pagination } from '@core/model/pagination.model';
 import { environment } from '@env';
 import { MainAspectType } from '@page/parts/model/mainAspectType.enum';
-import { Owner } from '@page/parts/model/owner.enum';
 import {
   AssetAsBuiltFilter,
   AssetAsPlannedFilter,
@@ -55,8 +54,8 @@ export class PartsService {
 
     let params = new HttpParams()
       .set('page', page)
-      .set('size', pageSize)
-      .set('filter', 'owner,EQUAL,OWN,AND');
+      .set('size', pageSize);
+
     sort.forEach(sortingItem => {
       params = params.append('sort', sortingItem);
     });
@@ -76,8 +75,7 @@ export class PartsService {
     let filterOperator = isOrSearch ? 'OR' : 'AND';
     let params = new HttpParams()
       .set('page', page)
-      .set('size', pageSize)
-      .set('filter', 'owner,EQUAL,OWN,AND');
+      .set('size', pageSize);
 
     sort.forEach(sortingItem => {
       params = params.append('sort', sortingItem);
@@ -93,7 +91,7 @@ export class PartsService {
   }
 
   public getPart(id: string): Observable<Part> {
-    if(!id || typeof id !== 'string') {
+    if (!id || typeof id !== 'string') {
       throw new Error('invalid ID');
     }
 
@@ -101,17 +99,17 @@ export class PartsService {
 
     const resultsAsBuilt = this.apiService.get<PartResponse>(`${ this.url }/assets/as-built/${ encodedId }`).pipe(
       map(part => PartsAssembler.assemblePart(part, MainAspectType.AS_BUILT)),
-      catchError(() => of(null))
+      catchError(() => of(null)),
     );
     const resultsAsPlanned = this.apiService.get<PartResponse>(`${ this.url }/assets/as-planned/${ encodedId }`).pipe(
       map(part => PartsAssembler.assemblePart(part, MainAspectType.AS_PLANNED)),
-      catchError(() =>  of(null))
+      catchError(() => of(null)),
     );
 
     // Combine both observables and filter out null values from the array
-    return forkJoin([resultsAsBuilt, resultsAsPlanned]).pipe(
-      filter(([partAsBuilt, partAsPlanned]) => partAsBuilt !== null || partAsPlanned !== null),
-      map(([partAsBuilt, partAsPlanned]) => partAsBuilt || partAsPlanned)
+    return forkJoin([ resultsAsBuilt, resultsAsPlanned ]).pipe(
+      filter(([ partAsBuilt, partAsPlanned ]) => partAsBuilt !== null || partAsPlanned !== null),
+      map(([ partAsBuilt, partAsPlanned ]) => partAsBuilt || partAsPlanned),
     );
   }
 
@@ -136,13 +134,12 @@ export class PartsService {
 
   }
 
-  public getDistinctFilterValues(isAsBuilt: boolean, owner: Owner, fieldNames: string, startsWith: string) {
+  public getDistinctFilterValues(isAsBuilt: boolean, fieldNames: string, startsWith: string) {
     const mappedFieldName = PartsAssembler.mapFieldNameToApi(fieldNames);
     let params = new HttpParams()
       .set('fieldName', mappedFieldName)
       .set('startWith', startsWith)
-      .set('size', 200)
-      .set('owner', owner);
+      .set('size', 200);
 
 
     if (isAsBuilt) {