+++ tx-backend/src/main/java/org/eclipse/tractusx/traceability/common/properties/BpdmProperties.java
@@ -1,5 +1,5 @@
 /********************************************************************************
- * Copyright (c) 2023 Contributors to the Eclipse Foundation
+ * Copyright (c) 2024 Contributors to the Eclipse Foundation
  *
  * See the NOTICE file(s) distributed with this work for additional
  * information regarding copyright ownership.
@@ -16,22 +16,25 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  ********************************************************************************/
-import { TableFilterConfiguration } from '@shared/components/parts-table/parts-config.model';
+package org.eclipse.tractusx.traceability.common.properties;
 
-export class PartsAsPlannedSupplierConfigurationModel extends TableFilterConfiguration {
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Configuration;
 
-  constructor() {
-    const sortableColumns = {
-      select: false,
-      semanticDataModel: true,
-      nameAtManufacturer: true,
-      businessPartner: true,
-      manufacturerName: true,
-      manufacturerPartId: true,
-      semanticModelId: true,
-      menu: false,
-    };
-    super(sortableColumns);
-  }
+@RequiredArgsConstructor
+@Getter
+@Configuration
+public class BpdmProperties {
+
+    @Value("${bpdm.bpnEndpoint}")
+    private String bpnEndpoint;
+    @Value("${bpdm.oAuthClientId}")
+    private String oAuthClientId;
+    @Value("${bpdm.timeout.read}")
+    private String readTimeout;
+    @Value("${bpdm.timeout.connect}")
+    private String connectTimeout;
+
 }
-